feat: Book Get API ErrorBoundary 적용
