+++ frontend/src/components/FallbackUI.tsx
@@ -29,6 +29,8 @@ export default FallbackUI;
 
 const FallbackUIContainer = css`
   position: relative;
+  width: 100%;
+  height: 100%;
   display: flex;
   flex-direction: column;
   justify-content: center;
